graph, gaphql: Move the herd cache to the graph crate
