+++ graph/Cargo.toml
@@ -11,6 +11,7 @@ atomic_refcell = "0.1.13"
 bigdecimal = { version = "0.1.0", features = ["serde"] }
 bytes = "1.0.1"
 cid = "0.11.0"
+defer = "0.1"
 diesel = { version = "1.4.8", features = ["postgres", "serde_json", "numeric", "r2d2", "chrono"] }
 diesel_derives = "1.4"
 chrono = "0.4.31"
